Feat: Drupal 9.3.x compatiblity.
file_url_transform_relative() is deprecated in Drupal 9.3.x in favour of the new
file_url_generator service.  To comply with this, we are removing all code
needed for file_url_transform_relative() to generate absolute URLs.  We have
replaced this with a service override for the new file_url_generator service
This overridden service generates absolute URLs even for relative URL requests
only when on the ModernGov template page.

@see https://www.drupal.org/node/2940031